unnecessary cast

+++ src/snes_apu/timer.rs
@@ -36,7 +36,7 @@ impl Timer {
 
             self.counter_low += 1;
             if let Some(target) = self.target {
-                if self.counter_low == target as u8 {
+                if self.counter_low == target {
                     self.counter_high += 1;
                     self.counter_low = 0;
                 }
